requirements.txt
jax==0.3.1
jaxlib==0.3.0
numpy==1.22.3
librosa==0.9.1
pax3==0.5.6
gradio
jinja2
toml==0.10.2
unidecode==1.3.4
pyyaml==6.0